Attempts at optimizations and catchup
Attempted to improve the performance a little bit, using both eivindbohler's pull request for the main repo and a Swift script edit that slows down how quickly the window renders, which Activity Monitor does claim lowers the CPU by half. Performance spikes are still a major problem, though, even if the average load is lower....

Also implemented a fix for the screen issue I've had: turned out the lack of interpolation when halving the LCD size was the issue.

Finally, implemented some minor but more recent changes from the main repo, most notably the removal of the battery warning and the implementation of the cursor.